Chrome breaks when trying to download file with comma in name
On our local nb install, we had a user complain that they couldn't download a file in Chrome. The user had a filename with commas in it - an odd choice, but should be supported. They were getting this error:

ERR_RESPONSE_HEADERS_MULTIPLE_CONTENT_DISPOSITION

This lead to a bug report, see: https://code.google.com/p/chromium/issues/detail?id=103618

By adding quotes around the filename, commas are supported. Additionally, the replacement of spaces with underscores might not be necessary (I didn't test that).