Fix server error that occurs when userinfo cookie is not valid JSON
We have been encountering issues with students receiving a Page Unavailable
message (500 error) after clicking on a bookmarklet that adds NB_embed.js to a
page to allow commenting on HTML (e.g., OpenStax textbooks). This error has been
found to occur after the bookmarklet is clicked while viewing NB Desktop, which
alters the userinfo cookie such that it is invalid JSON (because the JSON is
escaped twice -- once in buildEmbed.js and once again in the
`GLOB.auth.set_cookie` method -- cf fix in buildEmbed.js) even when decoded from
URL encoded format.

The issue does not appear to persist between browser sessions because the
userinfo cookie indicating that the user is a guest expires after the session
closes. However, clicking on the bookmarklet before logging in can also prevent
the user from logging in.